Solve Problem 3 - Largest Prime Factor
This one took way too long...

The MAXPRIMEFN is the maximum number of Prime factors that a
uint_64 can hold. (2^64) I wanted to refactor the code to have a
function that'd return a list of prime factors of a given prime but it's
not garuanteed to be an int64.

Perhaps a union with all possible int types?

A better program would ask for an input instead of having a hardcoded
value.