Solve Problem 4 - Largest Palindrome Product
This can absolutely be optimized as it's quite bruteforce-y.

If it is safe to assume that a product of two x-digits will always yield
a palindrome of at most 2*x digits, then there could be a "if k is less
than 2*x digits" to avoid the is_palindrome calculations.

Another optimization would be to check the potential multipliers of the
digits. If a palindrome starts with 9, it ends with 9. To get a number
that ends with 9, you need to multiply 9*1 (or 1*9), 3*3, or 7*7.

The first one would be useful at any x-digit, as opposed to the second
one, which would start saving a lot of computational time at higher
x-digits.