Solve Problem 5 - Smallest Multiple
This one is funny. I did it on paper in 2 minutes. First tried
multiplying all of the prime factors of all numbers in the range and got
a number too big for that to be the answer, then realized that if you
can divide by 10, you can by 2 and 5 so no need to multiply by 10.

I then tried going down the path of primes, their factors, and their
occurences. But it was quickly getting out of hand. I then redid it on
paper going from i = 2 to i = 10 and realized that the gcd would
simplify the whole thing.

I then looked online for a good C gcd implementation and saw a wiki
article using the lcm. So I used that as well.

~I cheated, yes.. fight me~